Add functions for accessing properties of frames
PFrame class implements several functions to access/read the flags of
it. VPage got a function to read the exclusive-flag from its properties.
PMemory can now add a bunch of frames represented by an iterator pair
representing the start iterator and end iterator of a range of
uint64_ts. Each such uint64_t is the number of a frame that should be
added. Like this the function can add frames independently from the
underlying container storing the frame numbers to add.